added libray vehicles

+++ app/Http/Controllers/API/UserController.php
@@ -5,12 +5,13 @@ namespace App\Http\Controllers\API;
 use App\Http\Controllers\API\BaseController as BaseController;
 use App\Http\Resources\ProfileResource;
 use App\Http\Resources\UserCollection;
+use App\Http\Resources\UserResource;
 use App\Profile;
 use App\Role;
 use App\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Validator;
-use App\Http\Resources\UserResource;
 
 class UserController extends BaseController
 {
@@ -25,10 +26,11 @@ class UserController extends BaseController
 
         return new UserCollection($users);
     }
+
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
@@ -36,7 +38,7 @@ class UserController extends BaseController
         $validator = Validator::make($request->all(), [
             'name' => 'required',
             'email' => 'required|email|unique:users,email',
-            'phone' => 'required|unique:users,phone',
+            'phone' => 'required',
             'role' => 'required|in:admin,staff,subscriber',
             'password' => 'required',
             'password_confirmation' => 'required|same:password',
@@ -59,8 +61,8 @@ class UserController extends BaseController
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param $id
+     * @return \Illuminate\Http\JsonResponse
      */
     public function show($id)
     {
@@ -76,37 +78,66 @@ class UserController extends BaseController
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param Request $request
+     * @param Profile $profile
+     * @return \Illuminate\Http\JsonResponse
      */
     public function update(Request $request, Profile $profile)
     {
         $input = $request->all();
 
         $validator = Validator::make($input, [
-            'notes' => 'required',
+            'bio' => 'required',
         ]);
 
-        if($validator->fails()){
+        if ($validator->fails()) {
             return $this->sendError('Validation Error.', $validator->errors());
         }
 
-        $profile->notes = $input['notes'];
+        $profile->bio = $input['bio'];
         $profile->save();
 
         return $this->sendResponse(new UserResource($profile), 'Profile updated successfully.');
     }
 
     /**
-     * Remove the specified resource from storage.
-     * @param User $user
-     * @return \Illuminate\Http\Response
-     * @throws \Exception
+     * @OA\Delete(
+     *      path="/api/v1/users/{id}",
+     *      operationId="destroy-user",
+     *      tags={"Profile"},
+     *      summary="Delete user.",
+     *      description="User has been deleted successfully.",
+     *      @OA\Parameter(
+     *          name="Authorization",
+     *          description="Bearer token",
+     *          required=true,
+     *          in="header",
+     *          @OA\Schema(type="string"),
+     *      ),
+     *      @OA\Response(
+     *          response=200,
+     *          description="User has been deleted successfully.",
+     *          @OA\MediaType(
+     *              mediaType="application/json",
+     *              @OA\JsonContent(type="object",example = {"success":true,"data":"","message":"User has been deleted successfully."})
+     *          )
+     *      ),
+     *      @OA\Response(
+     *          response=403,
+     *          description="Failed to delete.",
+     *          @OA\MediaType(
+     *              mediaType="application/json",
+     *              @OA\JsonContent(type="object",example = {"success":false,"message":"Failed to delete."})
+     *          )
+     *      ),
+     * )
      */
     public function destroy(User $user)
     {
-        $user->delete();
-        return $this->sendResponse([], 'User has been deleted successfully.');
+        if (Auth::id() != $user->id) {
+            $user->delete();
+            return $this->sendResponse([], 'User has been deleted successfully.');
+        }
+        return $this->sendError('Failed to delete.', [], 403);
     }
 }
